+++ CHANGELOG.md
@@ -98,7 +98,7 @@
 - Fix typo's.
 - Adjust tox matrix to currently supported Django versions.
 - Fixed names of tests.
-- Add tests for ``BaseRenderer``.
+- Add tests for ``IconRenderer``.
 
 ## 0.0.4 - 2017-06-22